+++ crates/cheatnet/src/cheatcodes/warp.rs
@@ -1,22 +1,17 @@
-use crate::{cheatcodes::EnhancedHintError, CheatedState};
+use crate::CheatnetState;
 use cairo_felt::Felt252;
 use starknet_api::core::ContractAddress;
 
-impl CheatedState {
-    pub fn start_warp(
-        &mut self,
-        contract_address: ContractAddress,
-        timestamp: Felt252,
-    ) -> Result<(), EnhancedHintError> {
-        self.warped_contracts.insert(contract_address, timestamp);
-        Ok(())
+impl CheatnetState {
+    pub fn start_warp(&mut self, contract_address: ContractAddress, timestamp: Felt252) {
+        self.cheatcode_state
+            .warped_contracts
+            .insert(contract_address, timestamp);
     }
 
-    pub fn stop_warp(
-        &mut self,
-        contract_address: ContractAddress,
-    ) -> Result<(), EnhancedHintError> {
-        self.warped_contracts.remove(&contract_address);
-        Ok(())
+    pub fn stop_warp(&mut self, contract_address: ContractAddress) {
+        self.cheatcode_state
+            .warped_contracts
+            .remove(&contract_address);
     }
 }