+++ CHANGELOG.md
@@ -14,8 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `#[should_panic]` attribute support
 - Documentation to public methods
 - Information sections to documentation about importing `snforge_std`
-- Added print support for basic numeric data types
->>>>>>> origin/master
+- Print support for basic numeric data types
+- Functions `parse_txt` and `TxtParser<T>::deserialize_txt` to load data from plain text files and serialize it
+- `get_class_hash` cheatcode
+- `mock_call` cheatcode
 
 #### Changed
 
@@ -47,6 +49,7 @@ snforge_std = { git = "https://github.com/foundry-rs/starknet-foundry", tag = "0
     - `compilation_unit_for_package`
 
 - Fails test when user tries to use syscalls not supported by forge test runner
+- Updated cairo-lang to 2.1.0, starknet-api to 0.4.1 and blockifier to 0.2.0-rc0
 
 ### Cast